Refactors window classes to inherit from Janela
This commit refactors several window classes to inherit from a base class named `Janela`.
This change promotes code reuse and provides a consistent look and feel across the application's windows.
It also allows to set the title of the windows more easily, since it is passed to the super constructor.